TweetStockSense/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
yfinance==0.2.33
tweepy==4.15.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.12.2
